Docs: 객체 별 모듈 분리
모델, 뷰, 컨트롤러를 개별 파일로 분리하고 main에 import 시키도록 코드 변경.

(cherry picked from commit 99ae7c44f08499281733c84f00581e124e389056)
